RCT example is updated with the bootstrap part. Now it can be run on Stata.
