--- java/MainEstimation/NaiveContainer.java
@@ -1,48 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2018 Stephen P. Ryan <stephen.p.ryan@wustl.edu>.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package MainEstimation;
-
-/**
- *
- * @author Stephen P. Ryan <stephen.p.ryan@wustl.edu>
- */
-public class NaiveContainer {
-
-    double MSPE_OLS;
-    int countSignificant;
-    
-    public NaiveContainer(int countSignificant, double MSPE_OLS) {
-        this.countSignificant = countSignificant;
-        this.MSPE_OLS = MSPE_OLS;
-    }
-
-    public int getCountSignificant() {
-        return countSignificant;
-    }
-
-    public double getMSPE_OLS() {
-        return MSPE_OLS;
-    }
-    
-}
