+++ java/core/readme.txt
@@ -1,0 +1,1 @@
+Directory for core files that will be used with any application.
