+++ java/examples/RCT/MomentContinuousSplitObjRCT.java
@@ -21,82 +21,89 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package MainEstimation;
+package examples.RCT;
 
-import Jama.Matrix;
-import MainEstimation.IntegerPartition;
-import MainEstimation.MomentPartitionObj;
-import MainEstimation.SplitContainer;
+import core.MomentContinuousSplitObj;
+import core.SplitContainer;
 
 /**
  *
  * @author Stephen P. Ryan <stephen.p.ryan@wustl.edu>
  */
-public class MomentPartitionObjBartRCT extends MomentPartitionObj {
+public class MomentContinuousSplitObjRCT extends MomentContinuousSplitObj {
 
     SplitContainer container;
-    
-    public MomentPartitionObjBartRCT(IntegerPartition partition, int indexSplitVariable, Matrix X, Matrix Y) {
-        this.partition = partition;
+    Jama.Matrix X;
+    Jama.Matrix Y;
+    int minCount;
+    double minProportion;
+
+    public MomentContinuousSplitObjRCT(int indexSplitVariable, Jama.Matrix X, Jama.Matrix Y, double minProportion, int minCount) {
         this.indexSplitVariable = indexSplitVariable;
         this.X = X;
         this.Y = Y;
-        
-        initialize();
+        this.minCount = minCount;
+        this.minProportion = minProportion;
+
+        // System.out.println("MomentContinuousSplitObjBart.java:26 -> min/max x_1 = "+pmUtility.min(X, 1)+" "+pmUtility.max(X,1));
     }
-    
-    private void initialize() {
-        container = getDataSplit();
-        numObsLeft = container.getyLeft().getRowDimension();
-        numObsRight = container.getyRight().getRowDimension();
-    }
-    
+
     @Override
     public int getNumObsLeft() {
         // in the rct context, care about the minimum of count of 0's and 1's in each partition
         int count = 0;
-        for (int i = 0; i < numObsLeft; i++) {
+        for (int i = 0; i < container.getxLeft().getRowDimension(); i++) {
             if (container.getxLeft().get(i, 0) == 0) {
                 count++;
             }
         }
-        return Math.min(count, numObsLeft - count);
+        return Math.min(count, container.getxLeft().getRowDimension() - count);
     }
 
     @Override
     public int getNumObsRight() {
         int count = 0;
-        for (int i = 0; i < numObsRight; i++) {
+        for (int i = 0; i < container.getxRight().getRowDimension(); i++) {
             if (container.getxRight().get(i, 0) == 0) {
                 count++;
             }
         }
-        return Math.min(count, numObsRight - count);
+        return Math.min(count, container.getxRight().getRowDimension() - count);
     }
-    
+
     @Override
     public double getMSE() {
+
         leftMSE = 0;
         rightMSE = 0;
 
         /**
          * RCT regresses outcome on indicator for treatment; there are no X's
-         * There is a constant, and we measure the coefficient on the W
-         * So in this implementation just use the first column to get OLS fits and errors, etc.
+         * There is a constant, and we measure the coefficient on the W So in
+         * this implementation just use the first column to get OLS fits and
+         * errors, etc.
          */
         ContainerRCT leftRCT = new ContainerRCT(container.getxLeft(), container.getyLeft());
         ContainerRCT rightRCT = new ContainerRCT(container.getxRight(), container.getyRight());
-        
+
         leftMSE = leftRCT.getMSE();
         rightMSE = rightRCT.getMSE();
-        
+
+        if (getNumObsLeft() < minCount || getNumObsRight() < minCount) {
+            return Double.POSITIVE_INFINITY;
+        }
+
         // System.out.println(numObsLeft+" "+numObsRight+" "+leftMSE+" "+rightMSE);
         // return (leftMSE + rightMSE) / X.getRowDimension();
         return (leftMSE + rightMSE);
     }
-    
-    
-    
-    
-    
+
+    @Override
+    public double f_to_minimize(double splitPoint) {
+        container = SplitContainer.getContinuousDataSplit(Y, X, splitPoint, indexSplitVariable);
+        numObsLeft = container.getyLeft().getRowDimension();
+        numObsRight = container.getyRight().getRowDimension();
+        return getMSE();
+    }
+
 }
